feat(ClassroomList):add validation on class delete

+++ src/components/ClassroomCard/ClassroomCard.tsx
@@ -1,7 +1,9 @@
 import Card from "@mui/material/Card";
 import CardContent from "@mui/material/CardContent";
+import MuiAlert, { AlertProps } from "@mui/material/Alert";
 
 import Grid from "@mui/material/Grid";
+import Snackbar from "@mui/material/Snackbar";
 import React, { FC, useState } from "react";
 import { classService } from "../../services/class.service";
 import {
@@ -12,6 +14,13 @@ import {
 import { useAppDispatch, useAppSelector } from "../../store.toolkit/store";
 import { RemoveModal } from "../RemoveModal/RemoveModal";
 
+const Alert = React.forwardRef<HTMLDivElement, AlertProps>(function Alert(
+  props,
+  ref
+) {
+  return <MuiAlert elevation={6} ref={ref} variant="filled" {...props} />;
+});
+
 export const ClassroomCard: FC<{ classroom: Classroom }> = ({ classroom }) => {
   const dispatch = useAppDispatch();
   const students = useAppSelector((state) => state.student.students);
@@ -19,6 +28,19 @@ export const ClassroomCard: FC<{ classroom: Classroom }> = ({ classroom }) => {
   const [open, setOpen] = useState<boolean>(false);
   const handleOpen = () => setOpen(true);
   const handleClose = () => setOpen(false);
+
+  const [openAlert, setOpenAlert] = useState(false);
+
+  const handleCloseAlert = (
+    event?: React.SyntheticEvent | Event,
+    reason?: string
+  ) => {
+    if (reason === "clickaway") {
+      return;
+    }
+
+    setOpenAlert(false);
+  };
 
   const deleteFromClass = (payload: {
     classroomId: number | string;
@@ -28,6 +50,11 @@ export const ClassroomCard: FC<{ classroom: Classroom }> = ({ classroom }) => {
   };
 
   const deleteClass = async () => {
+    if (classroom.students.length > 0) {
+      setOpenAlert(true);
+      return;
+    }
+
     try {
       classService.remove(classroom._id);
       dispatch(deleteClassroom(classroom._id));
@@ -58,6 +85,19 @@ export const ClassroomCard: FC<{ classroom: Classroom }> = ({ classroom }) => {
         students={students}
         deleteFromClass={deleteFromClass}
       />
+      <Snackbar
+        open={openAlert}
+        autoHideDuration={4000}
+        onClose={handleCloseAlert}
+      >
+        <Alert
+          onClose={handleCloseAlert}
+          severity="error"
+          sx={{ width: "100%" }}
+        >
+          Can't delete Class with student(s)
+        </Alert>
+      </Snackbar>
     </>
   );
 };
